# openapi/v1.yaml
openapi: 3.1.0
info:
  title: CenterSeat API
  version: 1.0.0
  description: |
    Contract for discovering screenings, fetching live seat inventory, and
    returning explainable best-seat recommendations. Query endpoints never hold
    or purchase seats.
servers:
  - url: http://localhost:8080
    description: Local Docker service
paths:
  /v1/movie-suggestions:
    get:
      operationId: suggestMovies
      summary: Resolve free-text movie input to canonical provider titles
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 160
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 6
      responses:
        "200":
          description: Ranked canonical movie matches from the configured source
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                type: object
                required: [suggestions]
                properties:
                  suggestions:
                    type: array
                    items:
                      $ref: "#/components/schemas/MovieSuggestion"
        "400":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/showtime-queries:
    post:
      operationId: createShowtimeQuery
      summary: Discover real screenings when live per-seat inventory is unavailable
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeatQueryRequest"
      responses:
        "200":
          description: Live showtime discovery results without seat claims
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShowtimeQueryResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "410":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/seat-queries:
    post:
      operationId: createSeatQuery
      summary: Find the best available seats across matching screenings
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeatQueryRequest"
      responses:
        "200":
          description: A fresh, ranked recommendation
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeatQueryResponse"
        "400":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "410":
          $ref: "#/components/responses/Problem"
        "422":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
  /v1/seat-queries/{query_id}:
    get:
      operationId: getSeatQuery
      summary: Read a previously completed query
      parameters:
        - name: query_id
          in: path
          required: true
          schema:
            type: string
        - name: If-None-Match
          in: header
          schema:
            type: string
      responses:
        "200":
          description: Cached query result
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Cache-Control:
              schema:
                type: string
              description: Private, short-lived result caching policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeatQueryResponse"
        "304":
          description: The client already has the current representation
        "404":
          $ref: "#/components/responses/Problem"
        "410":
          $ref: "#/components/responses/Problem"
  /v1/seat-queries/{query_id}/recommendations/{rank}:
    get:
      operationId: getSeatRecommendation
      summary: Refresh one recommendation and return its live seat map on demand
      parameters:
        - name: query_id
          in: path
          required: true
          schema:
            type: string
        - name: rank
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 5
      responses:
        "200":
          description: Fresh recommendation detail for the selected screening
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recommendation"
        "404":
          $ref: "#/components/responses/Problem"
        "409":
          $ref: "#/components/responses/Problem"
        "410":
          $ref: "#/components/responses/Problem"
        "503":
          $ref: "#/components/responses/Problem"
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Requests with the same key and body coalesce while in flight; later replays reuse the first result.
      schema:
        type: string
        minLength: 8
        maxLength: 128
  headers:
    ETag:
      schema:
        type: string
      description: Strong entity tag for the response body.
    RequestID:
      schema:
        type: string
      description: Correlation ID used in structured logs.
  responses:
    Problem:
      description: RFC 9457 problem detail
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    SeatQueryRequest:
      type: object
      additionalProperties: false
      required: [movie_query, location, dates, ticket_count, seat_profile, time]
      properties:
        movie_query:
          type: string
          minLength: 1
          maxLength: 160
          examples: [Dune]
        movie_id:
          type: string
          maxLength: 128
          description: Optional canonical provider movie identifier selected from movie suggestions.
        location:
          $ref: "#/components/schemas/LocationConstraint"
        dates:
          $ref: "#/components/schemas/DateConstraint"
        time:
          $ref: "#/components/schemas/TimeConstraint"
        ticket_count:
          type: integer
          minimum: 1
          maximum: 12
        seat_profile:
          type: string
          enum: [balanced, dead_center, two_thirds_back, aisle, front, back, custom]
        custom_seat_zone:
          allOf:
            - $ref: "#/components/schemas/SeatZone"
          description: Required when seat_profile is custom. Normalized auditorium bounds, where x runs left to right and y runs from the screen to the back row.
        formats:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [standard, imax, dolby, xd, rpdx, screenx, 3d, drive_in]
        language:
          type: string
        captions:
          type: string
          enum: [any, open, closed, none]
          default: any
        audio_description:
          type: boolean
          default: false
        wheelchair_spaces:
          type: integer
          minimum: 0
          maximum: 6
          default: 0
        companion_seats:
          type: integer
          minimum: 0
          maximum: 6
          default: 0
        amenities_required:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [recliner, dine_in, alcohol, laser_projection, reserved_seating]
        max_distance_miles:
          type: number
          exclusiveMinimum: 0
          maximum: 49
          default: 25
        max_total_price:
          type: number
          minimum: 0
        min_start_notice_minutes:
          type: integer
          minimum: 0
          maximum: 1440
          default: 0
          description: Screenings starting within this many minutes are excluded. Screenings that have already started are always excluded.
        allow_split_party:
          type: boolean
          default: false
        allow_unknown_price:
          type: boolean
          default: true
        exclude_first_rows:
          type: integer
          minimum: 0
          maximum: 10
          default: 0
        minimum_geometry_confidence:
          type: string
          enum: [exact_coordinates, rendered_geometry, row_geometry, label_heuristic]
          default: row_geometry
        candidate_limit:
          type: integer
          minimum: 1
          maximum: 64
          deprecated: true
          description: Compatibility-only upper bound. Interactive clients should omit this; the service compares every represented date and adaptively expands until the range winner is proven or the full automatic safety ceiling is reached.
    LocationConstraint:
      type: object
      additionalProperties: false
      required: [radius_miles]
      description: Supply latitude and longitude together, or a provider-supported coarse query such as a US ZIP code.
      properties:
        query:
          type: string
          description: City, postal code, neighborhood, or address.
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        radius_miles:
          type: number
          exclusiveMinimum: 0
          maximum: 49
    DateConstraint:
      type: object
      additionalProperties: false
      required: [start, end]
      properties:
        start:
          type: string
          format: date
        end:
          type: string
          format: date
    TimeConstraint:
      type: object
      additionalProperties: false
      required: [timezone]
      description: Times use the required IANA query timezone.
      properties:
        mode:
          type: string
          enum: [any, inside, outside, before, after]
          default: any
          description: For inside and outside, if start is later than end, the interval wraps past midnight; outside is its complement.
        start:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          description: Optional local time in HH:MM.
        end:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          description: Optional local time in HH:MM.
        timezone:
          type: string
          description: Required IANA time zone used to interpret query times and display screening times.
          examples: [America/New_York]
    SeatQueryResponse:
      type: object
      required: [query_id, status, generated_at, expires_at, refresh_until, coverage, winner, alternatives]
      properties:
        query_id:
          type: string
        status:
          type: string
          enum: [complete, partial, no_match]
        generated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
          description: Snapshot expiry for reading this result; distinct from the 15-minute refresh retention deadline.
        refresh_until:
          type: string
          format: date-time
          description: Retention deadline for refreshing recommendations; later requests return 410.
        coverage:
          $ref: "#/components/schemas/Coverage"
        winner:
          oneOf:
            - $ref: "#/components/schemas/Recommendation"
            - type: "null"
        alternatives:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"
        warnings:
          type: array
          items:
            type: string
...
    ShowtimeQueryResponse:
      type: object
      required: [query_id, status, generated_at, expires_at, coverage, showtimes]
      properties:
        query_id:
          type: string
        status:
          type: string
          enum: [complete, no_match]
        generated_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        coverage:
          $ref: "#/components/schemas/Coverage"
        showtimes:
          type: array
          items:
            $ref: "#/components/schemas/Showtime"
        warnings:
          type: array
          items:
            type: string
    Recommendation:
      type: object
      required:
        - rank
        - showtime
        - seats
        - score
        - confidence
        - explanation
        - profile_match
        - verified_at
        - seat_map
      properties:
        rank:
          type: integer
        showtime:
          $ref: "#/components/schemas/Showtime"
        seats:
          type: array
          items:
            $ref: "#/components/schemas/Seat"
        seat_options:
          type: array
          description: Strong currently available choices for a single-seat query, ordered with the primary recommendation first.
          items:
            $ref: "#/components/schemas/Seat"
        score:
          type: number
          minimum: 0
          maximum: 100
          description: Higher is better.
        confidence:
          type: string
          enum: [exact_coordinates, rendered_geometry, row_geometry, label_heuristic, visual_inference]
        explanation:
          type: array
          items:
            type: string
        score_breakdown:
          type: object
          additionalProperties:
            type: number
        profile_match:
          type: string
          enum: [preferred_zone, closest_fallback]
          description: Whether the recommendation is inside the selected profile's preferred viewing zone or is an explicitly labeled fallback.
        verified_at:
          type: string
          format: date-time
        booking_url:
          type: string
          format: uri
        seat_map:
          oneOf:
            - $ref: "#/components/schemas/SeatMap"
            - type: "null"
    Showtime:
      type: object
      required: [id, movie_title, venue_name, starts_at, format, distance_miles]
      properties:
        id:
          type: string
        movie_title:
          type: string
        venue_name:
          type: string
        auditorium_name:
          type: string
        starts_at:
          type: string
          format: date-time
        format:
          type: string
        distance_miles:
          type: number
        venue_timezone:
          type: string
          description: Optional IANA venue time zone used to interpret provider wall-clock screening times.
        total_price:
          type: number
        currency:
          type: string
    Seat:
      type: object
      required: [id, label, row, x, y, type, status]
      properties:
        id:
          type: string
        label:
          type: string
        row:
          type: string
        x:
          type: number
          minimum: 0
          maximum: 1
        y:
          type: number
          minimum: 0
          maximum: 1
        type:
          type: string
          enum: [standard, recliner, wheelchair, companion, sofa]
        status:
          type: string
          enum: [available, sold, held, broken, house]
    SeatMap:
      type: object
      required: [seats, target, confidence, observed_at]
      properties:
        seats:
          type: array
          items:
            $ref: "#/components/schemas/Seat"
        target:
          $ref: "#/components/schemas/GeometryPoint"
        preferred_depth:
          $ref: "#/components/schemas/DepthRange"
        preferred_zone_bounds:
          allOf:
            - $ref: "#/components/schemas/SeatZone"
          description: The normalized custom area applied to this auditorium, when a custom seat profile was requested.
        recommended_zone:
          type: array
          description: Up to four physical seat positions closest to the profile target, including unavailable positions.
          items:
            type: string
        confidence:
          type: string
          enum: [exact_coordinates, rendered_geometry, row_geometry, label_heuristic, visual_inference]
        observed_at:
          type: string
          format: date-time
...
    SeatZone:
      type: object
      additionalProperties: false
      required: [minimum_x, maximum_x, minimum_y, maximum_y]
      properties:
        minimum_x:
          type: number
          minimum: 0
          maximum: 1
        maximum_x:
          type: number
          minimum: 0
          maximum: 1
        minimum_y:
          type: number
          minimum: 0
          maximum: 1
          description: Front edge of the preferred area, measured from the screen.
        maximum_y:
          type: number
          minimum: 0
          maximum: 1
          description: Back edge of the preferred area.
...
    MovieSuggestion:
      type: object
      required: [id, title]
      properties:
        id:
          type: string
        title:
          type: string
        release_date:
          type: string
        year:
          type: string
    Problem:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
          enum: [validation_failed, invalid_request, not_found, query_expired, idempotency_key_expired, idempotency_conflict, screening_started, seats_unavailable, price_exceeded, provider_unavailable]
          description: Stable machine-readable error code.
        instance:
          type: string
          format: uri-reference